refactor home ui code
