+++ lib/features/home/presentation/widgets/home_screen_body.dart
@@ -1,4 +1,6 @@
 import 'package:ecommerce_website/core/shared/custom_footer_widget.dart';
+import 'package:ecommerce_website/features/home/presentation/widgets/flash_sales_header.dart';
+import 'package:ecommerce_website/features/home/presentation/widgets/product_grid.dart';
 import 'package:flutter/material.dart';
 
 class HomeScreenBody extends StatelessWidget {
@@ -21,220 +23,12 @@ class HomeScreenBody extends StatelessWidget {
             ),
           ),
           CustomFooterWidget()
-
         ],
       ),
     );
   }
 }
 
-class FlashSalesHeader extends StatelessWidget {
-  const FlashSalesHeader({super.key});
 
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.start,
-      children: [
-        Row(
-          children: [
-            Container(
-              width: 4,
-              height: 20,
-              color: Colors.red,
-            ),
-            const SizedBox(width: 8),
-            const Text(
-              "Our Products",
-              style: TextStyle(
-                color: Colors.red,
-                fontWeight: FontWeight.bold,
-                fontSize: 16,
-              ),
-            ),
-          ],
-        ),
-        const SizedBox(height: 8),
-        Row(
-          children: [
-            const Text(
-              'Premium Classic Watches',
-              style: TextStyle(
-                fontSize: 24,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-            const Spacer(),
-            const CountdownTimer(),
-            IconButton(
-              icon: const Icon(Icons.chevron_left),
-              onPressed: () {},
-            ),
-            IconButton(
-              icon: const Icon(Icons.chevron_right),
-              onPressed: () {},
-            ),
-          ],
-        ),
-      ],
-    );
-  }
-}
 
-class CountdownTimer extends StatelessWidget {
-  const CountdownTimer({super.key});
 
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        _buildTimeBox('03'),
-        const Text(' : '),
-        _buildTimeBox('23'),
-        const Text(' : '),
-        _buildTimeBox('19'),
-        const Text(' : '),
-        _buildTimeBox('56'),
-      ],
-    );
-  }
-
-  Widget _buildTimeBox(String time) {
-    return Container(
-      padding: const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-      decoration: BoxDecoration(
-        color: Colors.grey[200],
-        borderRadius: BorderRadius.circular(4),
-      ),
-      child: Text(
-        time,
-        style: const TextStyle(
-          fontWeight: FontWeight.bold,
-          fontSize: 16,
-        ),
-      ),
-    );
-  }
-}
-
-class ProductGrid extends StatelessWidget {
-  const ProductGrid({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return GridView.builder(
-      shrinkWrap: true,
-      physics: const NeverScrollableScrollPhysics(),
-      gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
-        crossAxisCount: 4,
-        childAspectRatio: 0.8,
-        crossAxisSpacing: 24,
-        mainAxisSpacing: 24,
-      ),
-      itemCount: 4,
-      itemBuilder: (context, index) {
-        return const ProductCard();
-      },
-    );
-  }
-}
-
-class ProductCard extends StatelessWidget {
-  const ProductCard({super.key});
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      elevation: 0,
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          Stack(
-            children: [
-              AspectRatio(
-                aspectRatio: 1,
-                child: Container(
-                  decoration: BoxDecoration(
-                    color: Colors.grey[200],
-                    borderRadius: BorderRadius.circular(8),
-                  ),
-                  child: const Center(
-                    child: Icon(Icons.image, size: 48, color: Colors.grey),
-                  ),
-                ),
-              ),
-              Positioned(
-                top: 8,
-                left: 8,
-                child: Container(
-                  padding:
-                      const EdgeInsets.symmetric(horizontal: 8, vertical: 4),
-                  decoration: BoxDecoration(
-                    color: Colors.red,
-                    borderRadius: BorderRadius.circular(4),
-                  ),
-                  child: const Text(
-                    '-40%',
-                    style: TextStyle(color: Colors.white, fontSize: 12),
-                  ),
-                ),
-              ),
-              const Positioned(
-                top: 8,
-                right: 8,
-                child: CircleAvatar(
-                  backgroundColor: Colors.white,
-                  radius: 16,
-                  child: Icon(Icons.favorite_border, size: 16),
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 8),
-          const Text(
-            'Gaming Controller',
-            style: TextStyle(
-              fontSize: 16,
-              fontWeight: FontWeight.bold,
-            ),
-          ),
-          const SizedBox(height: 4),
-          Row(
-            children: [
-              Text(
-                '\$80.00',
-                style: TextStyle(
-                  color: Colors.red[700],
-                  fontWeight: FontWeight.bold,
-                ),
-              ),
-              const SizedBox(width: 8),
-              Text(
-                '\$120.00',
-                style: TextStyle(
-                  color: Colors.grey[600],
-                  decoration: TextDecoration.lineThrough,
-                  fontSize: 12,
-                ),
-              ),
-            ],
-          ),
-          const SizedBox(height: 4),
-          Row(
-            children: [
-              const Icon(Icons.star, color: Colors.amber, size: 16),
-              const Text(
-                '4.8',
-                style: TextStyle(fontWeight: FontWeight.bold),
-              ),
-              Text(
-                ' (88)',
-                style: TextStyle(color: Colors.grey[600], fontSize: 12),
-              ),
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
