+++ lib/features/home/presentation/widgets/product_grid.dart
@@ -1,0 +1,24 @@
+import 'package:ecommerce_website/features/home/presentation/widgets/product_item.dart';
+import 'package:flutter/material.dart';
+
+class ProductGrid extends StatelessWidget {
+  const ProductGrid({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return GridView.builder(
+      shrinkWrap: true,
+      physics: const NeverScrollableScrollPhysics(),
+      gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
+        crossAxisCount: 4,
+        childAspectRatio: 0.8,
+        crossAxisSpacing: 24,
+        mainAxisSpacing: 24,
+      ),
+      itemCount: 4,
+      itemBuilder: (context, index) {
+        return const ProductCard();
+      },
+    );
+  }
+}
